Merge page and action variables
Until now, 'page' was meant for showing output in the default template and
'action' indicated what app's method to run (optional)

From now on, the 'page' param indicates the method to run. If it's not
found in the request, it's taken from the app's config variable 'init_page'.

The method now *is required to exists* in the app's class (and not be private).

The page that's used in the default template is, by default, that with the
same name than the contents of the 'page' param. It's stored in the variable
$app->page and so, can be changed in the method for different purpouses,
like redirection.