Change param errors behavior
From now, the validation errors will not be treated like a fatal
error (they still allow the app object to be created), but the errors
are kept into the app->errors array for the app to show them in the way
it thinks convenient.

Also, the params_error callback function is removed and instead, a new
redirection to the params_error_page key in the config will be done,
whenever this key is indicated.

All this, allows the app to treat the validation errors in a more
convenient way, and redirect the user automatically if said so by
configuration options