Don't use filter_var on our callback filters
The way filter_var calls the callback is too limited. The callback
only receives the value to check, and so, the result cannot be dependent
on other request parameters without scope tweaks.
Then, the callback only returns true or false, making it impossible to
return different errors or a combination of them.
This two problems, combined with the config only allowing one validator
per param, limits the use of all the validator framework much.

Using out own function call, we can send all the rest of request
params for validation testing, and return more than one posible error code.

As a side feature, some of the config errors that were simply making the
validation to stop, now get a proper error code and let the validation
continue.