Create a default logs folder if not specified
Create the logs folder from the logs class.
If not specified in the config, take PHP's temporary files folder; if
so, create it inside the one given.
This allows use to always have a sane default and is a first step to keep
it outside of the build folder without stupid Makefile rules.