Store the absolute path instead
This solves register_shutdown_function() not logging the errors with some
web servers.
In the called function, the working directory may have been
changed for the shutdown, so the relative path to logfile, is not correct anymore.
This will give you problems if you're using absolute paths in the config to start
with... but that's for another day.