Remove support to import the global app object
This means that the running app object is not available
for the templates at all.

If templates need to compute data, apps need to add it instead to
the templates with $this->template->set_var($name, $value). This means
that all business logic should be in the app, not in the template.

Apart from that, other variables are added automatically by the
parent class, such as params, errors and user info.